用户角色完善...

+++ kytc-user-server/src/main/java/com/kytc/user/server/impl/PermissionServiceImpl.java
@@ -1,12 +1,17 @@
 package com.kytc.user.server.impl;
 
+import com.kytc.framework.exception.BaseErrorCodeEnum;
+import com.kytc.framework.exception.BaseException;
 import com.kytc.framework.web.common.BasePageResponse;
 import com.kytc.framework.web.utils.BeanUtils;
+import com.kytc.user.request.PermissionSearchRequest;
 import com.kytc.user.server.service.PermissionService;
 import com.kytc.user.request.PermissionRequest;
 import com.kytc.user.response.PermissionResponse;
 import com.kytc.user.dao.data.PermissionData;
 import com.kytc.user.dao.mapper.PermissionMapperEx;
+import com.kytc.user.server.service.RolePermissionService;
+import com.kytc.user.server.service.UserPermissionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -19,16 +24,19 @@ import java.util.List;
 @RequiredArgsConstructor(onConstructor_={@Autowired})
 public class PermissionServiceImpl implements PermissionService {
 	private final PermissionMapperEx permissionMapperEx;
+	private final RolePermissionService rolePermissionService;
+	private final UserPermissionService userPermissionService;
 
 	@Override
-	public boolean add(PermissionRequest request){
-		if( null != request ){
-			PermissionData permissionData = BeanUtils.convert(request, PermissionData.class);
-			permissionData.setCreatedAt(new Date());
-			permissionData.setUpdatedAt(new Date());
-			return this.permissionMapperEx.insert(permissionData)>0;
+	public Long add(PermissionRequest request){
+		PermissionData permissionData = BeanUtils.convert(request, PermissionData.class);
+		permissionData.setCreatedAt(new Date());
+		permissionData.setUpdatedAt(new Date());
+		this.permissionMapperEx.insert(permissionData);
+		if( null == permissionData ){
+			throw new BaseException(BaseErrorCodeEnum.DATA_NOT_FOUND,"权限插入失败");
 		}
-		return false;
+		return permissionData.getId();
 	}
 
 	@Override
@@ -52,27 +60,34 @@ public class PermissionServiceImpl implements PermissionService {
 
 	@Override
 	public boolean delete(Long id){
+		if( this.rolePermissionService.countByPermissionId(id)>0 ){
+			throw new BaseException(BaseErrorCodeEnum.OPERATION_FAILED,"该权限已经关联角色,禁止删除");
+		}
+		if( this.userPermissionService.countByPermissionId(id)>0 ){
+			throw new BaseException(BaseErrorCodeEnum.OPERATION_FAILED,"该权限已经关联用户,禁止删除");
+		}
 		return this.permissionMapperEx.deleteByPrimaryKey(id)>0;
 	}
 
 	@Override
-	public BasePageResponse<PermissionResponse> listByCondition(PermissionRequest request,int page, int pageSize){
+	public BasePageResponse<PermissionResponse> listByCondition(PermissionSearchRequest request){
 		BasePageResponse<PermissionResponse> pageResponse = new BasePageResponse<>();
-		pageResponse.setRows(this.listByConditionData(request,page, pageSize));
+		pageResponse.setRows(this.listByConditionData(request));
 		pageResponse.setTotal(this.countByConditionData(request));
-		pageResponse.setPage(page);
-		pageResponse.setPageSize(pageSize);
+		pageResponse.setPage(request.getPage());
+		pageResponse.setPageSize(request.getPageSize());
 		return pageResponse;
 	}
 
-	private List<PermissionResponse> listByConditionData(PermissionRequest request,int page,int pageSize){
-		int start = page * pageSize;
-		List<PermissionData> list =  this.permissionMapperEx.listByCondition(request.getParentId(), request.getPermissionKey(), request.getPermissionName(), request.getUrl(), request.getLevel(), request.getDescription(), request.getIdDeleted(), start, pageSize);
+	private List<PermissionResponse> listByConditionData(PermissionSearchRequest request){
+		request.init();
+		List<PermissionData> list =  this.permissionMapperEx.listByCondition(request.getParentId(), request.getPermissionKey(),
+				request.getPermissionName(), request.getUrl(), request.getLevel(), request.getDescription(), request.getStart(), request.getLimit());
 		return BeanUtils.convert(list,PermissionResponse.class);
 	}
 
 
-	private Long countByConditionData(PermissionRequest request){
-		return this.permissionMapperEx.countByCondition(request.getParentId(), request.getPermissionKey(), request.getPermissionName(), request.getUrl(), request.getLevel(), request.getDescription(), request.getIdDeleted());
+	private Long countByConditionData(PermissionSearchRequest request){
+		return this.permissionMapperEx.countByCondition(request.getParentId(), request.getPermissionKey(), request.getPermissionName(), request.getUrl(), request.getLevel(), request.getDescription());
 	}
 }
